\name{isoSurv-package}
\alias{isoSurv-package}
\docType{package}
\title{Isotonic Regression on Survival Analysis}
\description{Nonparametric estimation on survival analysis under order restrictions}
\details{
\tabular{ll}{
Package: \tab isoph\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2020-3-20\cr
License: \tab GPL (>= 2)\cr
}
}
\author{Yunro Chung [aut,cre]
Maintainer: Yunro Chung <yunro.chung@asu.edu>
}
\references{Yunro Chung, Anastasia Ivanova, Michael G. Hudgens, Jason P. Fine (2018), Partial likelihood estimation of isotonic proportional hazards models, Biometrika, 105(1), 133-148. doi:10.1093/biomet/asx064}
\concept{Isotonic regression}
\concept{Nonparametric shape restricted estimation}
\concept{Survival analysis}
